feat: modify README.md, modify main to be able to read from and write to YAML file

+++ src/main/kotlin/StocksHistory.kt
@@ -3,5 +3,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class StocksHistory(
     val date: String,
-    val stocks: Map<String, Long>,
+    val stocks: Map<String, Int>,
 )
